moved not required headers to source files /  change codestyle

+++ src/promntline.c
@@ -1,11 +1,17 @@
 
-#include "../include/shell.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <termios.h>
+#include "../include/promntline.h"
 
 
-static void init_tty(struct termios* savtty){
+static void
+init_tty(struct termios *savtty) {
 
     struct termios tty;
-    if(tcgetattr(0, &tty)==-1){
+    if (tcgetattr(0, &tty) == -1) {
         perror("tcgetaattr fail");
         exit(1);
     }
@@ -13,7 +19,7 @@ static void init_tty(struct termios* savtty){
     tty.c_lflag &= ~(ECHO | ICANON);
     tty.c_cc[VMIN] = 1;
     tty.c_cc[VTIME] = 0;
-    if(tcsetattr(0, TCSAFLUSH, &tty)==-1){
+    if (tcsetattr(0, TCSAFLUSH, &tty) == -1) {
         perror("tcgetaattr fail");
         exit(1);
     }
@@ -21,47 +27,50 @@ static void init_tty(struct termios* savtty){
 }
 
 
-static void backspace(char* input_buff, size_t* curr, size_t* rpos){
+static void
+backspace(char *input_buff, size_t *curr, size_t *rpos) {
 
     --*curr;
     --*rpos;
-    for(size_t i=*curr; i < *rpos; ++i){
+    for (size_t i = *curr; i < *rpos; ++i) {
         input_buff[i] = input_buff[i + 1];
     }
 
     putchar('\010');
-    for(size_t i=*curr; i<*rpos; ++i){
+    for (size_t i = *curr; i < *rpos; ++i) {
         putchar(input_buff[i]);
     }
     putchar(' ');
-    for(size_t i=*curr; i<*rpos+1; ++i){
+    for (size_t i = *curr; i < *rpos + 1; ++i) {
         putchar('\010');
     }
 }
 
 
-static void insert_character(char* input_buff, size_t* curr, size_t* rpos, char ch){
+static void
+insert_character(char *input_buff, size_t *curr, size_t *rpos, char ch) {
 
-    for(size_t i = *rpos; i > *curr; --i){
+    for (size_t i = *rpos; i > *curr; --i) {
         input_buff[i] = input_buff[i - 1];
     }
     input_buff[*curr] = ch;
     ++*curr;
     ++*rpos;
 
-    for(size_t i=*curr-1; i<*rpos; ++i){
+    for (size_t i = *curr - 1; i < *rpos; ++i) {
         putchar(input_buff[i]);
     }
-    for(size_t i=*curr; i<*rpos; ++i){
+    for (size_t i = *curr; i < *rpos; ++i) {
         putchar('\010');
     }
 }
 
 
-
 #define MAX_SEQ_LEN 5
 #define MAX_INPUT_LEN 4096
-static int input_cmd(char* res, const char* original_cmd, int *is_up){
+
+static int
+input_cmd(char *res, const char *original_cmd, int *is_up) {
 
     static char input_buff[MAX_INPUT_LEN];
     static char read_buff[MAX_SEQ_LEN];
@@ -72,10 +81,10 @@ static int input_cmd(char* res, const char* original_cmd, int *is_up){
     printf("%s", input_buff);
     fflush(stdout);
 
-    while(1){
-        int cnt = read(0, read_buff, MAX_SEQ_LEN);
+    while (1) {
+        ssize_t cnt = read(0, read_buff, MAX_SEQ_LEN);
 
-        if(cnt==1){
+        if (cnt == 1) {
             switch (read_buff[0]) {
 
                 case 10:
@@ -85,10 +94,9 @@ static int input_cmd(char* res, const char* original_cmd, int *is_up){
                     return 1;
 
                 case 127:
-                    if(curr>0){
+                    if (curr > 0) {
                         backspace(input_buff, &curr, &rpos);
-                    }
-                    else{
+                    } else {
                         putchar('\a');
                     }
                     continue;
@@ -104,36 +112,34 @@ static int input_cmd(char* res, const char* original_cmd, int *is_up){
         }
 
 
-        if(cnt==3 && read_buff[0]=='\033' && read_buff[1]=='['){
+        if (cnt == 3 && read_buff[0] == '\033' && read_buff[1] == '[') {
 
-            switch(read_buff[2]) {
+            switch (read_buff[2]) {
                 case 'D':
-                    if(curr>0){
+                    if (curr > 0) {
                         --curr;
                         putchar('\010');
-                    }
-                    else{
+                    } else {
                         putchar('\a');
                     }
                     continue;
 
                 case 'C':
-                    if(curr < rpos){
+                    if (curr < rpos) {
                         ++curr;
                         printf("\033[1C");
                         fflush(stdout);
-                    }
-                    else{
+                    } else {
                         putchar('\a');
                     }
                     continue;
 
                 case 'A':
                 case 'B':
-                    for(int i=0; i<rpos-curr; ++i){
+                    for (int i = 0; i < rpos - curr; ++i) {
                         putchar(' ');
                     }
-                    for(int i=0; i<rpos; ++i){
+                    for (int i = 0; i < rpos; ++i) {
                         printf("\010 \010");
                         fflush(stdout);
                     }
@@ -145,11 +151,12 @@ static int input_cmd(char* res, const char* original_cmd, int *is_up){
 }
 
 
-void promptline(const char *prompt, char *line, vcharptr_t* history){
+void
+promptline(const char *prompt, char *line, vcharptr_t *history) {
 
     printf("\x1B[32m");
     fflush(stdout);
-    if(write(1, prompt, strlen(prompt))==-1){
+    if (write(1, prompt, strlen(prompt)) == -1) {
         perror("write fail");
         exit(1);
     }
@@ -161,31 +168,29 @@ void promptline(const char *prompt, char *line, vcharptr_t* history){
     int is_up;
 
     char empty[10] = "";
-    char* tmp_ptr = empty;
+    char *tmp_ptr = empty;
     vcharptr_t_push_back(history, &tmp_ptr);
 
-    int curr = (int)history->cnt - 1;
-    while(!input_cmd(line, history->ptr[curr], &is_up)){
-        if(is_up){
-            if(curr>0) {
+    int curr = (int) history->cnt - 1;
+    while (!input_cmd(line, history->ptr[curr], &is_up)) {
+        if (is_up) {
+            if (curr > 0) {
                 --curr;
-            }
-            else{
+            } else {
                 putchar('\a');
             }
             continue;
         }
-        if(curr < history->cnt - 1){
+        if (curr < history->cnt - 1) {
             ++curr;
-        }
-        else{
+        } else {
             putchar('\a');
         }
     }
     vcharptr_t_pop_back(history);
-    if(tcsetattr(0, TCSAFLUSH, &savetty)==-1){
+    if (tcsetattr(0, TCSAFLUSH, &savetty) == -1) {
         perror("tcsetattr fail");
         exit(1);
     }
- }
+}
